refactor: move 'named objects' into common
* NamedObject
* NamedJSONComparator

Removes a cyclic dependency: libAnki <-> AnkiDroid

Issue 18015
