+++ AnkiDroid/src/main/java/com/ichi2/anki/ui/NoteTypeSpinnerUtils.kt
@@ -20,8 +20,8 @@ import android.content.Context
 import android.widget.ArrayAdapter
 import android.widget.Spinner
 import com.ichi2.anki.R
+import com.ichi2.anki.common.json.NamedJSONComparator
 import com.ichi2.libanki.Collection
-import com.ichi2.utils.NamedJSONComparator
 
 fun setupNoteTypeSpinner(
     context: Context,
