+++ common/build.gradle.kts
@@ -55,7 +55,8 @@ dependencies {
     implementation(libs.jakewharton.timber)
     testImplementation(libs.junit.jupiter)
     testImplementation(libs.junit.vintage.engine)
-    testRuntimeOnly(libs.junit.platform.launcher)
+    testImplementation(libs.hamcrest)
+    testImplementation(libs.junit.platform.launcher)
     androidTestImplementation(libs.androidx.test.junit)
     androidTestImplementation(libs.androidx.espresso.core)
 }
